pyproject.toml
[project]
name = "supymem"
version = "0.1.0"
description = "AI-powered collaborative knowledge agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "redis>=5.0.1",
    "qdrant-client[fastembed]>=1.12.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "pgvector>=0.2.4",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "mem0ai>=0.1.0",
    "litellm>=1.55.0",
    "slack_bolt>=1.27.0",
    "PyGithub>=2.0.0",
    "PyJWT>=2.0.0",
    "cryptography>=42.0.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=24.1.0",
    "ruff>=0.1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
